chore: more golangci checks (#364)
* chore: fix a lot of godoc

and add a linter to catch them

* chore: avoid redefining min/max

And add a linter for avoiding such issues.

* chore: avoid unused parameter

The linter is still disabled as they are more errors to fix.
But it would require further checks and discussion, this is a good start

* chore: use common receiver names

Also enable a linter to detect such issues

* chore: enable most of the default revive rules

* chore: enable most gocritic rules

Apply the reported changes to the codebase.

Disable some rules that might be reporting false positives.
But it could also be bugs, so it requires manual review and discussion.